pyproject.toml
[project]
name = "pokemon-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pyyaml>=6.0.2",
]
